refactor: footer code cleanup
