+++ src/theme/Footer/Slope.jsx
@@ -1,0 +1,9 @@
+import React from 'react';
+
+import styles from './styles.module.css';
+
+const Slope = () => {
+  return <div className={styles.slope} />;
+};
+
+export default Slope;
